AYS-147 | Every method containing an object in the helper class has been moved to the relevant classes
