+++ src/test/java/org/ays/payload/RequestBodyUsers.java
@@ -13,4 +13,13 @@ public class RequestBodyUsers {
     private FiltersForUsers filter;
     private List<Sort> sort;
 
+    public static RequestBodyUsers generateCreateRequestBodyUsersWithPhoneNumberFilter(PhoneNumber phoneNumber) {
+        RequestBodyUsers requestBodyUsers = new RequestBodyUsers();
+        requestBodyUsers.setPagination(Pagination.createPagination());
+        FiltersForUsers filters = new FiltersForUsers();
+        filters.setPhoneNumber(phoneNumber);
+        requestBodyUsers.setFilter(filters);
+        return requestBodyUsers;
+    }
+
 }
