+++ src/test/java/org/ays/tests/auth/institutionAuthServiseTests/GetAdminTokenTest.java
@@ -3,14 +3,13 @@ package org.ays.tests.auth.institutionAuthServiseTests;
 import io.restassured.response.Response;
 import org.ays.endpoints.InstitutionAuthEndpoints;
 import org.ays.payload.AdminCredentials;
-import org.ays.payload.Helper;
 import org.ays.utility.AysResponseSpecs;
 import org.testng.annotations.Test;
 
 public class GetAdminTokenTest {
     @Test(groups = {"Smoke", "Regression", "Institution"})
     public void getTokenForValidAdmin() {
-        AdminCredentials adminCredentials = Helper.setIntsAdminCredentials();
+        AdminCredentials adminCredentials = AdminCredentials.generateIntsAdminCredentials();
         Response response = InstitutionAuthEndpoints.getAdminToken(adminCredentials);
         response.then()
                 .spec(AysResponseSpecs.expectSuccessResponseSpec())
