+++ src/test/java/org/ays/tests/auth/institutionAuthServiseTests/PostAdminInvalidateTokenTest.java
@@ -2,7 +2,7 @@ package org.ays.tests.auth.institutionAuthServiseTests;
 
 import io.restassured.response.Response;
 import org.ays.endpoints.InstitutionAuthEndpoints;
-import org.ays.payload.Helper;
+import org.ays.payload.AdminCredentials;
 import org.ays.payload.RefreshToken;
 import org.ays.payload.Token;
 import org.ays.utility.AysResponseSpecs;
@@ -11,7 +11,7 @@ import org.testng.annotations.Test;
 public class PostAdminInvalidateTokenTest {
     @Test(groups = {"Smoke", "Regression", "Institution"})
     public void adminInvalidateToken() {
-        Token token = Helper.getAdminToken(Helper.setIntsAdminCredentials());
+        Token token = Token.generateAdminToken(AdminCredentials.generateIntsAdminCredentials());
         RefreshToken refreshToken = new RefreshToken();
         refreshToken.setRefreshToken(token.getRefreshToken());
         Response response = InstitutionAuthEndpoints.adminInvalidateToken(token.getAccessToken(), refreshToken);
