+++ src/test/java/org/ays/tests/auth/institutionAuthServiseTests/PostAdminTokenRefreshTest.java
@@ -2,6 +2,7 @@ package org.ays.tests.auth.institutionAuthServiseTests;
 
 import io.restassured.response.Response;
 import org.ays.endpoints.InstitutionAuthEndpoints;
+import org.ays.payload.AdminCredentials;
 import org.ays.payload.Helper;
 import org.ays.payload.RefreshToken;
 import org.ays.payload.Token;
@@ -12,7 +13,7 @@ public class PostAdminTokenRefreshTest {
     @Test(groups = {"Smoke", "Regression", "Institution"})
     public void adminTokenRefresh() {
         RefreshToken refreshToken = new RefreshToken();
-        refreshToken.setRefreshToken(Helper.getAdminRefreshToken(Helper.setIntsAdminCredentials()));
+        refreshToken.setRefreshToken(Helper.getAdminRefreshToken(AdminCredentials.generateIntsAdminCredentials()));
         Response response = InstitutionAuthEndpoints.adminTokenRefresh(refreshToken);
         response.then()
                 .spec(AysResponseSpecs.expectSuccessResponseSpec())
@@ -20,7 +21,7 @@ public class PostAdminTokenRefreshTest {
     }
     @Test(groups = {"Regression", "Institution"})
     public void testAdminInvalidRefreshTokenForAccessTokenCreation() {
-        Token token = Helper.getAdminToken(Helper.setIntsAdminCredentials());
+        Token token = Token.generateAdminToken(AdminCredentials.generateIntsAdminCredentials());
         RefreshToken refreshToken = new RefreshToken();
         refreshToken.setRefreshToken(token.getRefreshToken());
         InstitutionAuthEndpoints.adminInvalidateToken(token.getAccessToken(), refreshToken);
