+++ src/test/java/org/ays/tests/user/assignmentmanagementservice/PostAssignmentStartTest.java
@@ -6,7 +6,6 @@ import org.ays.payload.Assignment;
 import org.ays.payload.Helper;
 import org.ays.payload.Location;
 import org.ays.payload.UserCredentials;
-import org.ays.utility.AysRandomUtil;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
